audio-processor/requirements.txt
# For librosa version (no ffmpeg required)
openai-whisper
librosa
soundfile
numpy
